Time: 0 ms (100%), Space: 7.8 MB (9.74%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -5,7 +5,7 @@ public:
         int end = num; 
         while(start <= end)
         {
-            int mid = start + (end - start) / 2;
+            long long mid = start + (end - start) / 2;
             if(mid * mid == num)
                 return 1;
             else if(mid * mid < num)
